feat: refactor public API into a class (#95)
## Which problem is this PR solving?
While working on the express example, @JamieDanielson realized that the
SDK didn't have a way to specify instrumentations. To fix this we
initially went down the path of [adding an `instrumentations` option to
`HoneycombOptions`](https://github.com/honeycombio/honeycomb-opentelemetry-node/pull/93)
so that we can configure instrumentation. After chatting about it some
more we realized that we've hit a use case for extending the `NodeSDK`
class instead of having a function that returns an instance of the
`NodeSDK` class. By extending the base OpenTelemetry `NodeSDK` class we
can support all the `NodeSDKConfiguration` options and build on top of
them for Honeycomb specific configuration. This not only solves our
`instrumentations` property problem but also future proofs our SDK from
having to make changes if the upstream configuration options change.

- Closes #92 

## Short description of the changes
- Adds `HoneycombSDK` class that extends `NodeSDK` and configures distro
specific options
- Removes `configureHoneycombSDK` function
- `HoneycombOptions` now extends `NodeSDKConfiguration` to support all
base options from the upstream SDK

## How to verify that this has the expected result
- [x] Spans are logged to the console when debug is true
- [x] Spans make it to Honeycomb when API key is set
- [x] Autoinstrumentation produces spans

Run the example app and curl the endpoint (`curl http://0.0.0.0:3000/`)
to see 5 total spans, including those from `instrumentationLibrary`s
for:

- manually created span (`hello-world-tracer`)
- automatic instrumentation (`@opentelemetry/instrumentation-http` in
the `hello-node` example,
- automatic instrumentation `@opentelemetry/instrumentation-http` and
`@opentelemetry/instrumentation-express` in the `hello-node-express`
example).

Note: With `debug` enabled, the first part of output will include the
instrumentations. This should probably be reconsidered in a separate
issue/PR, as it seems a bit noisy now.

Co-authored-by: Jamie Danielson <jamieedanielson@gmail.com>