feat: make metrics export configurable (#141)
## Which problem is this PR solving?

- Closes #138 

## Short description of the changes

- adds support for environment variables `OTEL_METRIC_EXPORT_INTERVAL`
and `OTEL_METRIC_EXPORT_TIMEOUT`
- uses provided environment variables to set `exportIntervalMillis` and
`exportTimeoutMillis` in metric exporter

Ideally these environment variables can/will be supported upstream, but
in the meantime this should allow end users to configure timeout and
interval behavior if needed. This is also used in our smoke tests to
reduce test time.

Note: Validation logic to ensure the metric interval is higher than the
metric timeout is handled upstream, and not included in this distro. If
the values provided are not positive numbers, we'll simply leave as
undefined (regular default behavior). The default values of 60000 and
30000 are set and used upstream.

## How to verify that this has the expected result

Set these environment variables to low numbers and run the example
express apps with debug to see more frequent export of metrics.

Co-authored-by: Purvi Kanal <purvikanal@honeycomb.io>