fix: add @opentelemetry/core as a dependency (#207)
## Which problem is this PR solving?
We need to add `@opentelemetry/core` as a dependency because we are
importing an interface and constant directly from it. I looked into
whether these values are re-exported from the trace sdk or api packages
and that doesn't seem to be the case. I think its okay for this package
to depend on `@opentelemetry/core`, other exporter packages have to
depend on it as well.

- Closes #206 

## Short description of the changes
- Added `@opentelemetry/core` to `package.json` as a core dependency.

## How to verify that this has the expected result
- `@opentelemetry/core` is listed is a direct dependency in
`package-lock.json`
- Existing tests pass
- Install works on versions of yarn and pnpm