fix: avoid dependency conflicts by using tilde for otel packages (#268)
## Which problem is this PR solving?

- Closes #264 

## Short description of the changes

- Experimental packages can have bigger (even breaking) changes in a
minor version, which makes a caret (minor) unpredictable. A tilde can be
used for allowing patches only.
- Some other changes seem to have issues sometimes as well (sdk-metrics
here) so setting all otel packages to tilde instead of caret. This will
need more frequent updates but it's better than breaking end user apps.

If the tilde is still too problematic, we'll need to consider pinning
versions in the future to avoid dependency conflicts with upstream.

I wasn't sure of a clean reliable way to check this in tests. I suppose
I can update the smoke tests to follow those steps I have in the below
repro section (building a tarball and using that as a dependency) but
wasn't sure if that made sense or to save that idea for later if this
keeps happening.

## How to verify that this has the expected result

**To reproduce error:**

On main
- open the hello-node-express-ts app
- change the distro dependency to be 0.6.0
(`"@honeycombio/opentelemetry-node": "^0.6.0",`)
- run `npm install` and `npm start`
- An error is thrown `TypeError: Cannot read properties of undefined
(reading 'explicitBucketBoundaries')`

**To see result of this change**

On this branch
- in the root directory run `npm i` and `npm pack` to create a tarball
- open the hello-node-express-ts app
- change the distro dependency to be that tarball
(`"@honeycombio/opentelemetry-node":
"/Users/jamiedanielson/code/honeycomb-opentelemetry-node/honeycombio-opentelemetry-node-0.6.0.tgz",`)
- run `npm install` and `npm start`
- See no error