feat: add support for classic ingest keys (#289)
## Which problem is this PR solving?

We've released Ingest Keys, but need to update the key detection logic
to allow Ingest Keys to be used to send data to Classic environments.

## Short description of the changes
- updates the Classic API Key detection logic to understand the shape of
a Classic Ingest Key
- updates tests with a more comprehensive list of cases
- bonus: updated the circleCI image for smoke tests to
`ubuntu-2204:2024.01.1` since the one we were using [is
deprecated](https://discuss.circleci.com/t/linux-image-deprecations-and-eol-for-2024/50177)

This is pretty much an exact copy of the logic in
https://github.com/honeycombio/honeycomb-opentelemetry-web/pull/101.

## How to verify that this has the expected result

It should be possible to use [an ingest
key](https://docs.honeycomb.io/working-with-your-data/settings/api-keys/#ingest-keys-1)
pointed to a classic environment after this change. Please note that the
ability to create ingest keys in classic environments isn't yet publicly
available (since we need to update this library and others to understand
classic ingest keys).