fix: ease restriction on api version (#299)
## Which problem is this PR solving?

- Follow-up from
https://github.com/honeycombio/honeycomb-opentelemetry-node/issues/293#issuecomment-2002328040

## Short description of the changes

- Use caret for versioning on API dependency, allowing for newer minor
versions

## How to verify that this has the expected result

Using this package `@honeycombio/opentelemetry-node` with upstream
packages should have minimal version conflicts. The more restrictive
version ranges (tilde instead of caret) were initially set because of
breaking changes across different packages that affected end users, and
the hope with the change was to minimize incompatibilities. However, the
API range may have been too restrictive - perhaps we can adjust this and
see if we need to ease up on others as well?